requirements.txt
mysql-connector-python==8.0.20
numpy==1.17.0
pandas==0.18.1
plotly==4.6.0
TA-Lib==0.4.18
urllib3==1.22
virtualenv==20.0.20
absl-py==0.9.0
aiohttp==3.6.2
# Editable install with no version control (alpha-vantage==2.1.3)
-e /usr/local/lib/python3.6/dist-packages
appdirs==1.4.4
apturl==0.5.2
asn1crypto==0.24.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
blis==0.4.1
Brlapi==0.6.6
cachetools==4.1.0
catalogue==1.0.0
certifi==2018.1.18
cffi==1.14.0
chardet==3.0.4
command-not-found==0.3
cryptography==2.9.2
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.3
defer==1.0.6
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz
f.lux-indicator-applet==1.1.11rc0
filelock==3.0.12
future==0.18.2
gast==0.3.3
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
httplib2==0.9.2
idna==2.6
idna-ssl==1.1.0
iexfinance==0.4.3
ImageHash==4.1.0
importlib-metadata==1.6.0
importlib-resources==1.5.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.2
MarkupSafe==1.0
matplotlib==3.2.1
mod-wsgi==4.7.1
more-itertools==8.3.0
mpl-finance==0.10.1
mplfinance==0.12.3a3
multidict==4.7.5
murmurhash==1.0.2
mysql-connector-python==8.0.20
ncclient==0.6.7
netifaces==0.10.4
numpy==1.17.0
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
opt-einsum==3.2.1
packaging==20.4
paho-mqtt==1.5.0
pandas==0.18.1
paramiko==2.7.1
pexpect==4.2.1
Pillow==5.1.0
plac==1.1.3
plotly==4.6.0
pluggy==0.13.1
preshed==3.0.2
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pytest==5.4.2
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
python-twitter==3.5
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
retrying==1.3.3
rsa==4.0
scipy==1.4.1
SecretStorage==2.3.1
selenium==3.141.0
simplejson==3.13.2
six==1.14.0
soupsieve==2.0
spacy==2.2.4
spacy-lookups-data==0.3.0
SQLAlchemy==1.3.17
srsly==1.0.2
ssh-import-id==5.7
system-service==0.3
systemd-python==234
TA-Lib==0.4.18
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
thinc==7.4.0
tqdm==4.46.0
typing-extensions==3.7.4.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==20.0.20
wadllib==1.3.2
wasabi==0.6.0
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
yarl==1.4.2
zipp==3.1.0
zope.interface==4.3.2
